# g7auto-be/openapi.yaml
openapi: 3.0.0
info:
  title: G7 Auto API
  description: Hệ thống quản lý showroom ô tô G7 Auto
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Đăng nhập
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh token
      responses:
        '200':
          description: OK
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Lấy thông tin người dùng hiện tại
      responses:
        '200':
          description: OK
  /api/auth/init-superadmin:
    post:
      tags: [Auth]
      summary: Khởi tạo tài khoản SUPERADMIN
      responses:
        '200':
          description: OK

  /api/accounts:
    get:
      tags: [Account]
      summary: Tìm kiếm tài khoản
      responses:
        '200':
          description: OK
    post:
      tags: [Account]
      summary: Tạo tài khoản mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '200':
          description: OK
  /api/accounts/{id}:
    get:
      tags: [Account]
      summary: Lấy chi tiết tài khoản
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    put:
      tags: [Account]
      summary: Cập nhật tài khoản
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountRequest'
      responses:
        '200':
          description: OK
    delete:
      tags: [Account]
      summary: Xóa tài khoản
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK

  /api/accounts_approving/search:
    get:
      tags: [Account Approval]
      summary: Tìm kiếm phê duyệt tài khoản
      responses:
        '200':
          description: OK
  /api/accounts_approving/search-pending:
    get:
      tags: [Account Approval]
      summary: Danh sách tài khoản chờ phê duyệt
      responses:
        '200':
          description: OK
  /api/accounts_approving/search-approved:
    get:
      tags: [Account Approval]
      summary: Danh sách tài khoản đã phê duyệt
      responses:
        '200':
          description: OK
  /api/accounts_approving/request-approval:
    post:
      tags: [Account Approval]
      summary: Gửi yêu cầu phê duyệt
      responses:
        '200':
          description: OK

  /api/employees:
    get:
      tags: [Employee]
      summary: Danh sách nhân viên
      responses:
        '200':
          description: OK
    post:
      tags: [Employee]
      summary: Tạo nhân viên mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeRequest'
      responses:
        '200':
          description: OK
  /api/employees/{id}:
    get:
      tags: [Employee]
      summary: Chi tiết nhân viên
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    put:
      tags: [Employee]
      summary: Cập nhật nhân viên
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeRequest'
      responses:
        '200':
          description: OK

  /api/employees-approving:
    get:
      tags: [Employee Approval]
      summary: Danh sách yêu cầu phê duyệt nhân viên
      responses:
        '200':
          description: OK
    post:
      tags: [Employee Approval]
      summary: Tạo yêu cầu phê duyệt mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeRequest'
      responses:
        '200':
          description: OK
  /api/employees-approving/{id}:
    put:
      tags: [Employee Approval]
      summary: Cập nhật yêu cầu phê duyệt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeRequest'
      responses:
        '200':
          description: OK
    delete:
      tags: [Employee Approval]
      summary: Xóa yêu cầu phê duyệt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK

  /api/cars:
    get:
      tags: [Car]
      summary: Tìm kiếm xe
      responses:
        '200':
          description: OK
    post:
      tags: [Car]
      summary: Tạo xe mới
      responses:
        '200':
          description: OK
  /api/cars/{id}:
    get:
      tags: [Car]
      summary: Lấy chi tiết xe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    put:
      tags: [Car]
      summary: Cập nhật xe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    delete:
      tags: [Car]
      summary: Xóa xe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK

  /api/car-models:
    get:
      tags: [CarModel]
      summary: Danh sách mẫu xe
      responses:
        '200':
          description: OK
    post:
      tags: [CarModel]
      summary: Tạo mẫu xe mới
      responses:
        '200':
          description: OK
  /api/car-models/{id}:
    get:
      tags: [CarModel]
      summary: Chi tiết mẫu xe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    put:
      tags: [CarModel]
      summary: Cập nhật mẫu xe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    delete:
      tags: [CarModel]
      summary: Xóa mẫu xe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK

  /api/showrooms:
    get:
      tags: [Showroom]
      summary: Danh sách showroom
      responses:
        '200':
          description: OK
    post:
      tags: [Showroom]
      summary: Tạo showroom mới
      responses:
        '200':
          description: OK
  /api/showrooms/list:
    get:
      tags: [Showroom]
      summary: Lấy danh sách showroom (không phân trang)
      responses:
        '200':
          description: OK
  /api/showrooms/{id}:
    get:
      tags: [Showroom]
      summary: Chi tiết showroom
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    put:
      tags: [Showroom]
      summary: Cập nhật showroom
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    delete:
      tags: [Showroom]
      summary: Xóa showroom
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK

  /api/customers:
    get:
      tags: [Customer]
      summary: Danh sách khách hàng
      responses:
        '200':
          description: OK
    post:
      tags: [Customer]
      summary: Tạo khách hàng mới
      responses:
        '200':
          description: OK
  /api/customers/{id}:
    get:
      tags: [Customer]
      summary: Chi tiết khách hàng
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    put:
      tags: [Customer]
      summary: Cập nhật khách hàng
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    delete:
      tags: [Customer]
      summary: Xóa khách hàng
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK

  /api/quotations:
    get:
      tags: [Quotation]
      summary: Danh sách báo giá
      responses:
        '200':
          description: OK
    post:
      tags: [Quotation]
      summary: Tạo báo giá mới
      responses:
        '200':
          description: OK
  /api/quotations/{id}:
    get:
      tags: [Quotation]
      summary: Chi tiết báo giá
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  /api/quotations/{id}/send:
    post:
      tags: [Quotation]
      summary: Gửi báo giá cho khách hàng
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK

  /api/deposits:
    get:
      tags: [Deposit]
      summary: Danh sách phiếu đặt cọc
      responses:
        '200':
          description: OK
    post:
      tags: [Deposit]
      summary: Tạo phiếu đặt cọc mới
      responses:
        '200':
          description: OK

  /api/contracts:
    get:
      tags: [Contract]
      summary: Danh sách hợp đồng
      responses:
        '200':
          description: OK
    post:
      tags: [Contract]
      summary: Tạo hợp đồng mới
      responses:
        '200':
          description: OK

  /api/payments:
    post:
      tags: [Payment]
      summary: Tạo mới thanh toán
      responses:
        '200':
          description: OK

  /api/car-transfers:
    get:
      tags: [CarTransfer]
      summary: Danh sách yêu cầu điều chuyển xe
      responses:
        '200':
          description: OK
    post:
      tags: [CarTransfer]
      summary: Tạo yêu cầu điều chuyển mới
      responses:
        '200':
          description: OK

  /api/test-drives:
    get:
      tags: [TestDrive]
      summary: Danh sách lịch lái thử
      responses:
        '200':
          description: OK
    post:
      tags: [TestDrive]
      summary: Đăng ký lịch lái thử mới
      responses:
        '200':
          description: OK

  /api/service-histories:
    post:
      tags: [ServiceHistory]
      summary: Tạo mới lịch sử tương tác
      responses:
        '200':
          description: OK

components:
  schemas:
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          pattern: '^[a-zA-Z0-9]+$'
          description: Tài khoản đăng nhập
        password:
          type: string
          description: Mật khẩu đăng nhập
      required:
        - username
        - password
    
    AccountRequest:
      type: object
      properties:
        username:
          type: string
          pattern: '^[a-zA-Z0-9]+$'
          description: Tên đăng nhập
        password:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
        roles:
          type: array
          items:
            type: string
        status:
          type: string
      required:
        - username
        - email
        - fullName

    EmployeeRequest:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        address:
          type: string
        birthDate:
          type: string
          format: date
        gender:
          type: string
        nationalId:
          type: string
        joinDate:
          type: string
          format: date
        showroomId:
          type: integer
          format: int64
        username:
          type: string
          pattern: '^[a-zA-Z0-9]+$'
          description: Tên đăng nhập hệ thống tương ứng
      required:
        - fullName
        - phone
        - showroomId

    EmployeeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        birthDate:
          type: string
          format: date
        gender:
          type: string
        nationalId:
          type: string
        joinDate:
          type: string
          format: date
        employeeStatus:
          type: string
        showroomId:
          type: integer
          format: int64
        showroomName:
          type: string
        username:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string

    AccountApprovingResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        roles:
          type: string
        status:
          type: string
        statusApproving:
          type: string
        action:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        createdBy:
          type: string
        updatedBy:
          type: string

    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        fullName:
          type: string
        roles:
          type: array
          items:
            type: string
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        createdBy:
          type: string
        updatedBy:
          type: string